+++ core/routes/versions.ts
@@ -1,5 +1,6 @@
 import type { Context, Hono } from "../../deps/hono.ts";
 import type { CMSContent } from "../../types.ts";
+import { dispatch } from "../utils/event.ts";
 import { getPath } from "../utils/path.ts";
 
 interface Data {
@@ -24,23 +25,24 @@ export default function (app: Hono) {
     response.headers.set("X-Lume-CMS", "reload");
 
     if (action === "create") {
-      await versioning.create(name);
-      await versioning.change(name);
+      versioning.create(name);
+      versioning.change(name);
       return response;
     }
 
     if (action === "change") {
-      await versioning.change(name);
+      versioning.change(name);
       return response;
     }
 
     if (action === "publish") {
-      await versioning.publish(name);
+      versioning.publish(name);
+      dispatch("versionPublished", { name });
       return response;
     }
 
     if (action === "delete") {
-      await versioning.delete(name);
+      versioning.delete(name);
       return response;
     }
   });